tested changeset logic
